backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
deepeval==2.9.4
Deprecated==1.2.18
distro==1.9.0
execnet==2.1.1
firecrawl-py==2.6.0
Flask==3.1.1
flask-cors==6.0.0
frozenlist==1.6.0
google-auth==2.40.1
google-genai==1.16.1
googleapis-common-protos==1.70.0
greenlet==3.2.2
groq==0.25.0
grpcio==1.71.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-core==0.3.60
langchain-openai==0.3.17
langchain-pinecone==0.2.6
langchain-tests==0.3.19
langchain-text-splitters==0.3.8
langsmith==0.3.42
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
monotonic==1.6
multidict==6.4.4
nest-asyncio==1.6.0
numpy==2.2.6
ollama==0.4.8
openai==1.79.0
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
orjson==3.10.18
packaging==24.2
pinecone==6.0.2
pinecone-plugin-interface==0.0.7
playwright==1.52.0
pluggy==1.6.0
portalocker==3.1.1
posthog==3.25.0
propcache==0.3.1
protobuf==5.29.4
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.1
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-repeat==0.9.4
pytest-rerunfailures==12.0
pytest-socket==0.7.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9.1
sentry-sdk==2.29.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
svix==1.65.0
syrupy==4.9.1
tabulate==0.9.0
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
typer==0.15.4
types-Deprecated==1.2.15.20250304
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0
